añadido feeback para los metodos de borrado

+++ src/main/java/pes/anticatastrofe/aditionalInfo/AditionalInfoController.java
@@ -4,8 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pes.anticatastrofe.admin.Admin;
-import pes.anticatastrofe.admin.AdminDTO;
 
 import java.util.HashMap;
 import java.util.List;
@@ -13,7 +11,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping(path="/aditional_info")
+@RequestMapping(path = "/aditional_info")
 
 public class AditionalInfoController {
     private final AditionalInfoService aditionalInfoService;
@@ -27,21 +25,20 @@ public class AditionalInfoController {
     public List<AditionalInfoDTO> getAditionalInfos() {
         List<AditionalInfo> users = aditionalInfoService.getAditionalInfos();
         List<AditionalInfoDTO> aditionalInfoDTOS = users.stream()
-                .map(aditionalInfo->new AditionalInfoDTO(aditionalInfo))
+                .map(aditionalInfo -> new AditionalInfoDTO(aditionalInfo))
                 .collect(Collectors.toList());
         return aditionalInfoDTOS;
     }
 
     @PostMapping
-    public ResponseEntity<Map<String, String>> registerNewAditionalInfo(@RequestBody AditionalInfo aditionalInfo){
+    public ResponseEntity<Map<String, String>> registerNewAditionalInfo(@RequestBody AditionalInfo aditionalInfo) {
         Map<String, String> response = new HashMap<String, String>();
-        if(!aditionalInfoService.findByID(aditionalInfo.email).isPresent()) {
+        if (!aditionalInfoService.findByID(aditionalInfo.email).isPresent()) {
             AditionalInfo ai = aditionalInfoService.addNewAditionalInfo(aditionalInfo);
             response.put("operation_success", "true");
             response.put("deleted_object_id", ai.email);
             return new ResponseEntity<>(response, HttpStatus.OK);
-        }
-        else {
+        } else {
             response.put("message", "aditionalInfo already exists");
             response.put("status", HttpStatus.ALREADY_REPORTED.toString());
             return new ResponseEntity<>(response, HttpStatus.ALREADY_REPORTED);
@@ -49,11 +46,16 @@ public class AditionalInfoController {
     }
 
     @DeleteMapping
-    public Map<String,String> deleteAditionalInfo(@RequestParam String email) {
-        aditionalInfoService.deleteAditionalInfo(email);
-        Map<String, String> map = new HashMap<String, String>();
-        map.put("operation_success", "true");
-        map.put("deleted_person_id",email);
-        return map;
+    public ResponseEntity<Map<String, String>> deleteAditionalInfo(@RequestParam String email) {
+        Map<String, String> response = new HashMap<String, String>();
+        if (aditionalInfoService.findByID(email).isPresent()) {
+            aditionalInfoService.deleteAditionalInfo(email);
+            response.put("operation_success", "true");
+            return new ResponseEntity<>(response, HttpStatus.OK);
+        } else {
+            response.put("message", "aditionalInfo not exists so nothing was deleted");
+            response.put("status", HttpStatus.NOT_FOUND.toString());
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
     }
 }
