+++ src/main/java/pes/anticatastrofe/landmark/LandmarkController.java
@@ -12,13 +12,13 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping(path="/landmark")
+@RequestMapping(path = "/landmark")
 public class LandmarkController {
     private final LandmarkService landmarkService;
     private final TagService tagService;
 
     @Autowired
-    public LandmarkController(LandmarkService landmarkService,TagService tagService) {
+    public LandmarkController(LandmarkService landmarkService, TagService tagService) {
         this.landmarkService = landmarkService;
         this.tagService = tagService;
     }
@@ -27,13 +27,13 @@ public class LandmarkController {
     public List<LandmarkDTO> getLandmarks() {
         List<Landmark> landmarks = landmarkService.getLandmarks();
         List<LandmarkDTO> landmarksDTO = landmarks.stream()
-                .map(landmark->new LandmarkDTO(landmark))
+                .map(landmark -> new LandmarkDTO(landmark))
                 .collect(Collectors.toList());
         return landmarksDTO;
     }
 
     @PostMapping
-    public ResponseEntity<Map<String, String>> registerNewLandmark(@RequestBody Landmark landmark){
+    public ResponseEntity<Map<String, String>> registerNewLandmark(@RequestBody Landmark landmark) {
         Map<String, String> response = new HashMap<String, String>();
         if (tagService.getTagById(landmark.tag).isPresent()) {
             if (!landmarkService.getLandmarkById(landmark.id).isPresent()) {
@@ -41,14 +41,12 @@ public class LandmarkController {
                 response.put("operation_success", "true");
                 response.put("new_landmark_id", Integer.toString(l.id));
                 return new ResponseEntity<>(response, HttpStatus.OK);
-            }
-            else {
+            } else {
                 response.put("message", "landmark already exists");
                 response.put("status", HttpStatus.ALREADY_REPORTED.toString());
                 return new ResponseEntity<>(response, HttpStatus.ALREADY_REPORTED);
             }
-        }
-        else {
+        } else {
             response.put("message", "tag not exists");
             response.put("status", HttpStatus.FAILED_DEPENDENCY.toString());
             return new ResponseEntity<>(response, HttpStatus.FAILED_DEPENDENCY);
@@ -56,11 +54,17 @@ public class LandmarkController {
     }
 
     @DeleteMapping
-    public Map<String,String> deleteLandmark(@RequestParam Integer landmark_id) {
-        landmarkService.deleteLandmark(landmark_id);
-        Map<String, String> map = new HashMap<String, String>();
-        map.put("operation_success", "true");
-        map.put("deleted_landmark_id",Integer.toString(landmark_id));
-        return map;
+    public ResponseEntity<Map<String, String>> deleteLandmark(@RequestParam Integer landmark_id) {
+        Map<String, String> response = new HashMap<String, String>();
+        if (landmarkService.findByID(landmark_id).isPresent()) {
+            landmarkService.deleteLandmark(landmark_id);
+            response.put("operation_success", "true");
+            response.put("deleted_landmark_id", Integer.toString(landmark_id));
+            return new ResponseEntity<>(response, HttpStatus.OK);
+        } else {
+            response.put("message", "landmark not exists so nothing was deleted");
+            response.put("status", HttpStatus.NOT_FOUND.toString());
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
     }
 }
