+++ src/main/java/pes/anticatastrofe/tag/TagController.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pes.anticatastrofe.pin.Pin;
 
 import java.util.HashMap;
 import java.util.List;
@@ -12,7 +11,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping(path="/tag")
+@RequestMapping(path = "/tag")
 public class TagController {
     private final TagService tagService;
 
@@ -25,35 +24,40 @@ public class TagController {
     public List<TagDTO> getTags() {
         List<Tag> pins = tagService.getTags();
         List<TagDTO> tagsDTO = pins.stream()
-                .map(tag->new TagDTO(tag))
+                .map(tag -> new TagDTO(tag))
                 .collect(Collectors.toList());
         return tagsDTO;
     }
 
     @PostMapping
-    public ResponseEntity<Map<String, String>> registerNewTag(@RequestBody Tag tag){
+    public ResponseEntity<Map<String, String>> registerNewTag(@RequestBody Tag tag) {
         Map<String, String> response = new HashMap<String, String>();
-        if(!tagService.findByID(tag.name).isPresent()) {
+        if (!tagService.findByID(tag.name).isPresent()) {
             Tag t = tagService.addNewTag(tag);
             response.put("operation_success", "true");
-            response.put("new_tag_id",t.name);
+            response.put("new_tag_id", t.name);
             response.put("operation_success", "true");
             response.put("deleted_object_id", t.name);
             return new ResponseEntity<>(response, HttpStatus.OK);
-        }
-        else {
-            response.put("message", "aditionalInfo already exists");
+        } else {
+            response.put("message", "tag already exists");
             response.put("status", HttpStatus.ALREADY_REPORTED.toString());
             return new ResponseEntity<>(response, HttpStatus.ALREADY_REPORTED);
         }
     }
 
     @DeleteMapping
-    public Map<String,String> deleteTag(@RequestParam String name) {
-        tagService.deleteTag(name);
-        Map<String, String> map = new HashMap<String, String>();
-        map.put("operation_success", "true");
-        map.put("deleted_tag_id",name);
-        return map;
+    public ResponseEntity<Map<String, String>> deleteTag(@RequestParam String name) {
+        Map<String, String> response = new HashMap<String, String>();
+        if (tagService.findByID(name).isPresent()) {
+            tagService.deleteTag(name);
+            response.put("operation_success", "true");
+            response.put("deleted_tag_id", name);
+            return new ResponseEntity<>(response, HttpStatus.OK);
+        } else {
+            response.put("message", "tag not exists so nothing was deleted");
+            response.put("status", HttpStatus.NOT_FOUND.toString());
+            return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        }
     }
 }
